[#67] Add router directory & router for fileUpload
